requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.1.1
Werkzeug==3.0.1
alembic==1.13.1
